+++ test/cli.js
@@ -1,14 +1,10 @@
-'use strict'
+import { after, before, describe, it } from 'node:test'
+import fs from 'node:fs/promises'
+import path from 'node:path'
 
-const fs = require('node:fs/promises')
-const path = require('node:path')
-
-const access = require('./helpers/access')
-const describeInstaller = require('./helpers/describe_installer')
-const spawn = require('../src/spawn')
-
-const tempOutputDir = describeInstaller.tempOutputDir
-const cleanupOutputDir = describeInstaller.cleanupOutputDir
+import access from './helpers/access.js'
+import { cleanupOutputDir, tempOutputDir } from './helpers/describe_installer.js'
+import spawn from '../src/spawn.js'
 
 function runCLI (options) {
   const args = [
@@ -21,8 +17,7 @@ function runCLI (options) {
   before(() => spawn('./src/cli.js', args))
 }
 
-describe('cli', function () {
-  this.timeout(10000)
+describe('cli', () => {
 
   describe('with an app with asar', () => {
     const outputDir = tempOutputDir()